Extract route fetching config.
FetchRoutes is initialized explicitly in ApiRouter.init/1 because we
need dynamic configuration of the options. Specifically, we want to
fetch the routes from the Config process and the name of the
configuration can be user supplied so we cannot set it using
Plug.Builder pipelines as the options are set at compile time.