Add simple rules resource to API.
We cannot include Plug options in the response because we cannot
JSON-encode arbitrary terms.

In the future (?) there will be a richer configuration abstraction that
might allow us avoid that problem by outputting more domain-specific
data structures.

In any case, this resource is really only intended for introspection and
testing and isn't really a big deal.